+++ js/nodes/Node.js
@@ -327,7 +327,7 @@ define( function( require ) {
         
         var oldBounds = this._bounds;
         
-        var newBounds = this.isVisible() ? this.localToParentBounds( this._selfBounds ).union( that.localToParentBounds( this._childBounds ) ) : Bounds2.NOTHING;
+        var newBounds = this.localToParentBounds( this._selfBounds ).union( that.localToParentBounds( this._childBounds ) );
         var changed = !newBounds.equals( oldBounds );
         
         if ( changed ) {
@@ -357,7 +357,7 @@ define( function( require ) {
           var childBounds = Bounds2.NOTHING.copy();
           _.each( that.children, function( child ) { childBounds.includeBounds( child._bounds ); } );
           
-          var fullBounds = that.isVisible() ? that.localToParentBounds( that._selfBounds ).union( that.localToParentBounds( childBounds ) ) : Bounds2.NOTHING;
+          var fullBounds = that.localToParentBounds( that._selfBounds ).union( that.localToParentBounds( childBounds ) );
           
           sceneryAssertExtra && sceneryAssertExtra( that._childBounds.equalsEpsilon( childBounds, epsilon ), 'Child bounds mismatch after validateBounds: ' +
                                                                                                     that._childBounds.toString() + ', expected: ' + childBounds.toString() );
@@ -509,16 +509,18 @@ define( function( require ) {
       return this._bounds;
     },
     
-    // like getBounds() in the "parent" coordinate frame, but also includes invisible descendants
-    getCompleteBounds: function() {
+    // like getBounds() in the "parent" coordinate frame, but includes only visible nodes
+    getVisibleBounds: function() {
       // defensive copy, since we use mutable modifications below
       var bounds = this._selfBounds.copy();
       
       _.each( this.children, function( child ) {
-        bounds.includeBounds( child.getCompleteBounds() );
+        if ( child.isVisible() ) {
+          bounds.includeBounds( child.getVisibleBounds() );
+        }
       } );
       
-      sceneryAssert && sceneryAssert( bounds.isFinite() || bounds.isEmpty(), 'Complete bounds should not be infinite' );
+      sceneryAssert && sceneryAssert( bounds.isFinite() || bounds.isEmpty(), 'Visible bounds should not be infinite' );
       return this.localToParentBounds( bounds );
     },
     
@@ -970,7 +972,6 @@ define( function( require ) {
         
         this._visible = visible;
         
-        this.invalidateBounds(); // since visibility can affect bounds
         this.notifyVisibilityChange();
       }
       return this;
@@ -1663,7 +1664,7 @@ define( function( require ) {
     get selfBounds() { return this.getSelfBounds(); },
     get childBounds() { return this.getChildBounds(); },
     get globalBounds() { return this.getGlobalBounds(); },
-    get completeBounds() { return this.getCompleteBounds(); },
+    get visibleBounds() { return this.getVisibleBounds(); },
     get id() { return this.getId(); },
     get instances() { return this.getInstances(); },