+++ tests/qunit/js/scene.js
@@ -756,17 +756,17 @@
     color.setRGBA( 0, 255, 0, 1 );
   } );
   
-  test( 'Bounds and Complete Bounds', function() {
+  test( 'Bounds and Visible Bounds', function() {
     var node = new scenery.Node();
     var rect = new scenery.Rectangle( 0, 0, 100, 50 );
     node.addChild( rect );
     
-    ok( node.bounds.equals( new dot.Bounds2( 0, 0, 100, 50 ) ), 'Bounds Visible' );
-    ok( node.completeBounds.equals( new dot.Bounds2( 0, 0, 100, 50 ) ), 'Complete Bounds Visible' );
+    ok( node.visibleBounds.equals( new dot.Bounds2( 0, 0, 100, 50 ) ), 'Visible Bounds Visible' );
+    ok( node.bounds.equals( new dot.Bounds2( 0, 0, 100, 50 ) ), 'Complete Bounds Visible' );
     
     rect.visible = false;
     
-    ok( node.bounds.equals( dot.Bounds2.NOTHING ), 'Bounds Invisible' );
-    ok( node.completeBounds.equals( new dot.Bounds2( 0, 0, 100, 50 ) ), 'Complete Bounds Invisible' );
+    ok( node.visibleBounds.equals( dot.Bounds2.NOTHING ), 'Visible Bounds Invisible' );
+    ok( node.bounds.equals( new dot.Bounds2( 0, 0, 100, 50 ) ), 'Complete Bounds Invisible' );
   } );
 })();